[c-5717] Saved page lineup to tanstack query
