+++ packages/common/src/api/index.ts
@@ -25,6 +25,7 @@ export * from './tan-query/useFollowers'
 export * from './tan-query/useFollowing'
 export * from './tan-query/useHandleInUse'
 export * from './tan-query/useHandleReservedStatus'
+export * from './tan-query/useLibraryTracks'
 export * from './tan-query/useMutualFollowers'
 export * from './tan-query/usePurchasers'
 export * from './tan-query/usePurchasersCount'
