+++ packages/web/src/pages/saved-page/SavedPage.tsx
@@ -1,14 +1,26 @@
 import { useIsMobile } from 'hooks/useIsMobile'
 
-import SavedPageProvider from './SavedPageProvider'
 import DesktopSavedPage from './components/desktop/SavedPage'
 import MobileSavedPage from './components/mobile/SavedPage'
 
+const messages = {
+  title: 'Library',
+  description: "View tracks that you've favorited"
+}
+
 const SavedPage = () => {
   const isMobile = useIsMobile()
-  const content = isMobile ? MobileSavedPage : DesktopSavedPage
 
-  return <SavedPageProvider>{content}</SavedPageProvider>
+  const commonProps = {
+    title: messages.title,
+    description: messages.description
+  }
+
+  return isMobile ? (
+    <MobileSavedPage {...commonProps} />
+  ) : (
+    <DesktopSavedPage {...commonProps} />
+  )
 }
 
 export default SavedPage
