+++ packages/web/src/pages/saved-page/components/desktop/SavedPage.module.css
@@ -1,13 +1,32 @@
+/* Page Layout */
 .savedPageWrapper {
   position: relative;
   text-align: left;
 }
 
+.bodyWrapper {
+  display: flex;
+  flex-direction: column;
+  position: relative;
+}
+
+/* Header */
 .savedPageHeader {
   justify-content: left;
 }
 
-/* Filter Input in Header */
+.headerBottomBarContainer {
+  display: flex;
+  justify-content: space-between;
+  align-items: end;
+}
+
+/* Filter */
+.filterContainer {
+  transition: all 0.5s;
+  padding-bottom: var(--harmony-unit-5);
+}
+
 .savedPageWrapper :global(.ant-input-affix-wrapper) {
   position: absolute;
   right: 0px;
@@ -15,6 +34,7 @@
   width: 120px;
   z-index: 1;
 }
+
 .savedPageWrapper :global(.ant-input-affix-wrapper .ant-input) {
   border-radius: 4px;
   border: 1px solid var(--harmony-n-100) !important;
@@ -27,53 +47,45 @@
   font-size: var(--harmony-font-xs);
   font-weight: var(--harmony-font-medium);
   border: none;
-
   padding-top: 1px;
   padding-bottom: 1px;
   height: 24px;
 }
+
 .savedPageWrapper :global(.ant-input-prefix) > svg {
   width: 12px;
   height: 12px;
 }
+
 .savedPageWrapper :global(.ant-input-prefix) > svg g path {
   fill: var(--harmony-n-400);
 }
 
-.filterContainer {
-  transition: all 0.5s;
-  padding-bottom: var(--harmony-unit-5);
+/* Tabs */
+.tabElement {
+  justify-content: flex-start;
 }
 
-/* Body Content Wrapper */
-.bodyWrapper {
-  display: flex;
-  flex-direction: column;
-  position: relative;
+/* Loading */
+.spinner {
+  height: 40px;
+  width: 40px;
+  margin: 0 auto;
 }
 
-/* Track Table */
-.bodyWrapper .tableWrapper {
-  width: 100%;
-  border-radius: 8px;
-  box-shadow: 0 2px 5px -2px rgba(133, 129, 153, 0.25);
-  background-color: var(--harmony-white);
-  padding-bottom: 5px;
-  overflow: hidden;
+.spinner g path {
+  stroke: var(--harmony-secondary);
 }
 
-/* Albums */
+/* Cards */
 .cardsContainer {
   gap: var(--harmony-unit-6) var(--harmony-unit-4);
   margin-bottom: var(--harmony-unit-4);
 }
 
+/* Play Button */
 .playButtonContainer {
   transition: all 0.25s;
-}
-
-.tabElement {
-  justify-content: flex-start;
 }
 
 .albumsWrapper {
@@ -101,20 +113,4 @@
 
 .createPlaylistIcon circle {
   fill: var(--harmony-n-600);
-}
-
-.spinner {
-  height: 40px;
-  width: 40px;
-  margin: 0 auto;
-}
-
-.spinner g path {
-  stroke: var(--harmony-secondary);
-}
-
-.headerBottomBarContainer {
-  display: flex;
-  justify-content: space-between;
-  align-items: end;
 }
