+++ packages/web/src/pages/saved-page/components/desktop/TracksTabPage.module.css
@@ -1,0 +1,9 @@
+.container {
+  display: flex;
+  flex-direction: column;
+  gap: var(--unit-4);
+}
+
+.filterContainer {
+  width: 340px;
+}
