+++ packages/web/src/pages/saved-page/components/mobile/SavedPage.module.css
@@ -1,7 +1,13 @@
+/* Page Layout */
 .mobilePageContainer {
   height: 100%;
 }
 
+.pageContainer {
+  margin-top: 48px;
+}
+
+/* Header */
 .header {
   border-bottom: 1px solid var(--harmony-n-100);
   flex-wrap: wrap;
@@ -14,23 +20,6 @@
   width: 100%;
 }
 
-.spinner {
-  height: 40px;
-  width: 40px;
-  margin: var(--harmony-unit-18) auto 0;
-}
-
-.spinner g path {
-  stroke: var(--harmony-secondary) !important;
-}
-
-.tabContainer {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  margin-top: 40px;
-}
-
 .tabBar {
   position: fixed;
   top: 116px;
@@ -41,24 +30,15 @@
   z-index: 10;
 }
 
-.pageContainer {
-  margin-top: 48px;
-}
-
-.lineupContainer {
-  padding: 12px 11px 12px;
-}
-
-.tracksLineupContainer {
-  -webkit-overflow-scrolling: touch;
+/* Tab Container */
+.tabContainer {
+  display: flex;
+  flex-direction: column;
   width: 100%;
+  margin-top: 40px;
 }
 
-.cardLineupContainer {
-  width: 100%;
-}
-
-/* Search Container */
+/* Search */
 .searchContainer {
   height: 56px;
   margin: 12px;
@@ -111,6 +91,23 @@
   fill: var(--harmony-n-400);
 }
 
+/* Loading */
+.spinner {
+  height: 40px;
+  width: 40px;
+  margin: var(--harmony-unit-18) auto 0;
+}
+
+.spinner g path {
+  stroke: var(--harmony-secondary) !important;
+}
+
+/* Track List */
+.tracksLineupContainer {
+  -webkit-overflow-scrolling: touch;
+  width: 100%;
+}
+
 .trackListContainer {
   margin: 16px 12px;
   border: 1px solid var(--harmony-n-100);
@@ -123,13 +120,18 @@
     0 2px 5px -2px var(--tile-shadow-3);
 }
 
+/* Cards */
+.cardLineupContainer {
+  width: 100%;
+}
+
 .cardsContainer {
   gap: var(--harmony-unit-6) var(--harmony-unit-4);
   margin-top: 16px;
   padding: 0px 11px;
 }
 
-/* Empty Tab */
+/* Empty State */
 .emptyTab {
   margin: 126px 11px 8px;
   padding: 34px 24px 32px;
